pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cleanreport"
version = "0.0.1"
description = "Data quality & cleaning pipline with YAML configurable steps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas>=2.0.0", "numpy>=1.24.0", "pyyaml>=6.0"]

[project.scripts] 
cleanreport = "cleanreport.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["cleanreport"]

